extraindo classes do shape
