+++ src/test/java/edu/fatec/oo/ShapeTest.java
@@ -7,18 +7,6 @@ class ShapeTest {
 
     @Test
     public void test() {
-
-//        Shape s1 = new Shape("name1");
-//
-//        System.out.println(s1);
-//
-//        Shape s2 = new Shape();
-//
-//        System.out.println(s2);
-
-//        Square square1 = new Square();
-//        System.out.println(square1);
-
         Rectangle rectangle1 = new Rectangle(2.0, 3.0);
 
         System.out.println(rectangle1);
@@ -42,23 +30,21 @@ class ShapeTest {
         Shape shape3 = new Trapezoid(2.0,4.0,2.0, 3.0, 4.0);
         System.out.println(shape3);
         System.out.println("Area do shape3: " + shape3.area());
-
-
-
     }
 
-    @Test
-    void testExceptionsClassShape() {
-        assertThrows(RuntimeException.class, () -> new Shape("shape1").area());
-        assertThrows(RuntimeException.class, () -> new Shape("shape2").perimeter());
-        // testando as mensaagens das excecoes
-        assertEquals("not supported", assertThrows(RuntimeException.class, () -> new Shape("shape1").area()).getMessage());
-        assertEquals("not supported", assertThrows(RuntimeException.class, () -> new Shape("shape1").perimeter()).getMessage());
-
-    }
+//    @Test
+//    void testExceptionsClassShape() {
+//        assertThrows(RuntimeException.class, () -> new Shape("shape1").area());
+//        assertThrows(RuntimeException.class, () -> new Shape("shape2").perimeter());
+//        // testando as mensaagens das excecoes
+//        assertEquals("not supported", assertThrows(RuntimeException.class, () -> new Shape("shape1").area()).getMessage());
+//        assertEquals("not supported", assertThrows(RuntimeException.class, () -> new Shape("shape1").perimeter()).getMessage());
+//
+//    }
 
     @Test
     void testShapeString() {
         assertEquals("Shape.toString(): name=square", new Shape("square").toString());
     }
+
 }
